% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.r
\name{distE.diff}
\alias{distE.diff}
\title{Calculate pooled dispersion via MMAD distribution or median}
\usage{
distE.diff(x, y, mid = FALSE)
}
\arguments{
\item{x}{A vector of numbers for the first distribution}

\item{y}{A vector of numbers for the second distribution}

\item{mid}{Toggles whether to give the distribution or the median}
}
\description{
\code{distE.diff} calculates the difference vector of two distributions
}
