% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.r
\name{distE.effect}
\alias{distE.effect}
\title{Calculate nonparametric effect size vector}
\usage{
distE.effect(x, y, mid = TRUE)
}
\arguments{
\item{x}{A vector of numbers for the first distribution}

\item{y}{A vector of numbers for the second distribution}

\item{mid}{Toggles whether to give the distribution or the median}
}
\description{
\code{distE.effect} calculates the effect vector of two distributions
}
