% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmad.r
\name{distE.mmad}
\alias{distE.mmad}
\title{Calculate pooled dispersion via MMAD distribution or median}
\usage{
distE.mmad(x, y, mid = FALSE)
}
\arguments{
\item{x}{A vector of numbers for the first distribution}

\item{y}{A vector of numbers for the second distribution}

\item{mid}{Toggles whether to give the distribution or the median}
}
\description{
\code{distE.mmad} calculates the maximum absolute difference within
two groups and returns the dispersion vector as a result.
}
