% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.r
\name{distE.ol}
\alias{distE.ol}
\title{Calculate pooled dispersion via MMAD distribution or median}
\usage{
distE.ol(x)
}
\arguments{
\item{x}{A vector of effect sizes}
}
\description{
\code{distE.ol} calculates the overlap of distributions given the effect size vector
}
